+++ packages/vscode-extension/CHANGELOG.md
@@ -1,6 +1,8 @@
 ## 0.8.0
 
 - Removed support for moon < 1.0.
+- Refactored extension to support multiple VS Code workspace folders.
+	- When you open a file in another workspace, the moon panels will refresh.
 
 ## 0.7.0